+++ types/group_consumer.d.ts
@@ -1,15 +1,12 @@
-/// <reference path="./base_consumer.d.ts" />
-/// <reference path="./index.d.ts" />
 
-import Kafka = require("no-kafka");
-import { Client } from "client";
-import { BaseConsumer } from "base_consumer";
+
 
 declare module "group_consumer" {
+    import { BaseConsumer } from "base_consumer";
 
     export class GroupConsumer extends BaseConsumer {
-        commitOffset(commits: Commit[]): Promise<any>;
+        // commitOffset(commits: Commit[]): Promise<any>;
         fetchOffset(commits): Promise<number[]>;
-        
+
     }
 }
